Status Correction

Detect errors, or detect errors and correct, the tree life status evolution over the censuses. Inspired by the code of Nino Page package (ForestData::correct_alive() and .correct_alive_tree())

Usage

StatusCorrection(
  Data,
  InvariantColumns = c("Site", "Genus_TreeDataCor", "Species_TreeDataCor",
    "Family_TreeDataCor", "ScientificName_TreeDataCor"),
  DeathConfirmation = 2,
  UseSize = FALSE,
  AddRowsForForgottenCensuses = TRUE,
  DetectOnly = FALSE,
  RemoveRBeforeAlive = FALSE,
  RemoveRAfterDeath = FALSE
)

Arguments

Data

Dataset (data.frame or data.table) The LifeStatus column must be coded as: - TRUE = alive, - FALSE = dead, - NA = unseen The Plot column is needed to add rows to the census where the plot was inventoried, where the tree was alive, but not recorded.

InvariantColumns

Vector with the names of the columns that are supposed to have always the same value for each measurement of the same tree (character). It is recommended to use the columns that have already been corrected, such as the columns containing the corrected botanical information.

DeathConfirmation

Number of times (censuses) needed for an unseen tree to be considered dead (numeric) (Default = 2 censuses)

UseSize

Use the size presence as a witness of the living status of the tree (logical) (Default = FALSE)

AddRowsForForgottenCensuses

TRUE: adds rows for forgotten censuses between 2 'Alive', FALSE: does not add any rows (logical)

DetectOnly

TRUE: Only detect errors, FALSE: detect and correct errors (logical)

RemoveRBeforeAlive

Do you want to delete the rows about the tree before it was seen alive for the 1st time? (logical)

RemoveRAfterDeath

After correction do you want to delete the rows concerning the tree after its death? (logical)

Value

Fill the Comment column with error type informations. If DetectOnly = FALSE, add a LifeStatus_TreeDataCor column with the corrected trees life status.

Details

if UseSize : if Diameter != NA -> Alive

Dead > Alive -> NA

add rows for the forgotten censuses between 2 'Alive' if chosen

Alive > Dead/NA > Alive -> Alive

Alive > NA > Dead -> NA

Alive > Dead > NA -> Dead

Alive > NA > NA: if DeathConfirmation > unseens -> NA if DeathConfirmation =< unseens -> Dead

Examples

library(data.table)
data(TestData)

selection <- c("101184", "101433","101435","101436")

# Write the sequence
TestData <- TestData[order(Year)] # arrange years in ascending order
TestData[IdTree == "101184", LifeStatus := c(TRUE, TRUE, TRUE, TRUE, FALSE)]
#>          Site Plot Subplot PlotLat PlotLon XPlotUTM YPlotUTM PlotArea
#>    1: Paracou    6    none      NA      NA       NA       NA     6.25
#>    2: Paracou    6    none      NA      NA       NA       NA     6.25
#>    3: Paracou    6    none      NA      NA       NA       NA     6.25
#>    4: Paracou    6    none      NA      NA       NA       NA     6.25
#>    5: Paracou    6    none      NA      NA       NA       NA     6.25
#>   ---                                                                
#> 4851: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4852: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4853: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4854: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4855: Paracou    6    none      NA      NA       NA       NA     6.25
#>       PlotElevation SubplotLat SubplotLon XSubplotUTM YSubplotUTM SubplotArea
#>    1:            NA         NA         NA          NA          NA          NA
#>    2:            NA         NA         NA          NA          NA          NA
#>    3:            NA         NA         NA          NA          NA          NA
#>    4:            NA         NA         NA          NA          NA          NA
#>    5:            NA         NA         NA          NA          NA          NA
#>   ---                                                                        
#> 4851:            NA         NA         NA          NA          NA          NA
#> 4852:            NA         NA         NA          NA          NA          NA
#> 4853:            NA         NA         NA          NA          NA          NA
#> 4854:            NA         NA         NA          NA          NA          NA
#> 4855:            NA         NA         NA          NA          NA          NA
#>       MinDBH IdCensus Year Month  Day       Date TreeFieldNum IdTree
#>    1:  10.03     2016 2016  <NA> <NA> 2016-09-14           45 100665
#>    2:  10.03     2016 2016  <NA> <NA> 2016-09-14          148 100765
#>    3:  10.03     2016 2016  <NA> <NA> 2016-09-14           46 100666
#>    4:  10.03     2016 2016  <NA> <NA> 2016-09-14           49 100669
#>    5:  10.03     2016 2016  <NA> <NA> 2016-09-14            1 100621
#>   ---                                                               
#> 4851:  10.03     2020 2020  <NA> <NA> 2020-06-04          884 101479
#> 4852:  10.03     2020 2020  <NA> <NA> 2020-06-04          887 101482
#> 4853:  10.03     2020 2020  <NA> <NA> 2020-06-04          896 101491
#> 4854:  10.03     2020 2020  <NA> <NA> 2020-06-04          905 101500
#> 4855:  10.03     2020   NA  <NA> <NA> 2020-06-04          764 101362
#>       StemFieldNum        IdStem LifeStatus Diameter BD  Circ BCirc HOM POM
#>    1:         <NA> 100665_1_auto       TRUE    53.16 NA 167.0    NA  NA   0
#>    2:         <NA> 100765_1_auto       TRUE    14.32 NA  45.0    NA  NA   0
#>    3:         <NA> 100666_1_auto       TRUE    19.89 NA  62.5    NA  NA   0
#>    4:         <NA> 100669_1_auto       TRUE    41.38 NA 130.0    NA  NA   0
#>    5:         <NA> 100621_1_auto       TRUE    13.37 NA  42.0    NA  NA   0
#>   ---                                                                      
#> 4851:         <NA> 101479_1_auto       TRUE    16.39 NA  51.5    NA  NA   0
#> 4852:         <NA> 101482_1_auto       TRUE    20.85 NA  65.5    NA  NA   0
#> 4853:         <NA> 101491_1_auto       TRUE    29.60 NA  93.0    NA  NA   0
#> 4854:         <NA> 101500_1_auto       TRUE    27.53 NA  86.5    NA  NA   0
#> 4855:         <NA> 101362_1_auto       TRUE    27.85 NA  87.5    NA  NA   0
#>       BHOM BPOM TreeHeight TreeLat TreeLon XTreeUTM YTreeUTM XTreePlot
#>    1:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    2:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    3:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    4:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    5:   NA <NA>         NA      NA      NA       NA       NA        NA
#>   ---                                                                 
#> 4851:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4852:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4853:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4854:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4855:   NA <NA>         NA      NA      NA       NA       NA        NA
#>       YTreePlot XTreeSubplot YTreeSubplot        ScientificName   VernName
#>    1:        NA           NA           NA        Eperua falcata       wapa
#>    2:        NA           NA           NA Conceveiba guianensis     kusisi
#>    3:        NA           NA           NA    Moronobea coccinea  Moronobea
#>    4:        NA           NA           NA   Pouteria melanopoda    bakuman
#>    5:        NA           NA           NA Sandwithia guyanensis  wata tiki
#>   ---                                                                     
#> 4851:        NA           NA           NA   Lecythis persistens maho rouge
#> 4852:        NA           NA           NA   Lecythis persistens maho rouge
#> 4853:        NA           NA           NA        Eperua falcata       wapa
#> 4854:        NA           NA           NA          Licania alba       koko
#> 4855:        NA           NA           NA     Licania canescens       koko
#>                 Family      Genus    Species Subspecies Variety Voucher IdLevel
#>    1:         Fabaceae     Eperua    falcata       <NA>    <NA>    <NA>    <NA>
#>    2:    Euphorbiaceae Conceveiba guianensis       <NA>    <NA>    <NA>    <NA>
#>    3:       Clusiaceae  Moronobea   coccinea       <NA>    <NA>    <NA>    <NA>
#>    4:       Sapotaceae   Pouteria melanopoda       <NA>    <NA>    <NA>    <NA>
#>    5:    Euphorbiaceae Sandwithia guyanensis       <NA>    <NA>    <NA>    <NA>
#>   ---                                                                          
#> 4851:    Lecythidaceae   Lecythis persistens       <NA>    <NA>    <NA>    <NA>
#> 4852:    Lecythidaceae   Lecythis persistens       <NA>    <NA>    <NA>    <NA>
#> 4853:         Fabaceae     Eperua    falcata       <NA>    <NA>    <NA>    <NA>
#> 4854: Chrysobalanaceae    Licania       alba       <NA>    <NA>    <NA>    <NA>
#> 4855: Chrysobalanaceae    Licania  canescens       <NA>    <NA>    <NA>    <NA>
#>       Authority CommercialSp LifeForm LifeStatusOriginal CommercialSpOriginal
#>    1:      <NA>         TRUE     <NA>               TRUE                 TRUE
#>    2:      <NA>        FALSE     <NA>               TRUE                FALSE
#>    3:      <NA>         TRUE     <NA>               TRUE                 TRUE
#>    4:      <NA>        FALSE     <NA>               TRUE                FALSE
#>    5:      <NA>        FALSE     <NA>               TRUE                FALSE
#>   ---                                                                        
#> 4851:      <NA>        FALSE     <NA>               TRUE                FALSE
#> 4852:      <NA>        FALSE     <NA>               TRUE                FALSE
#> 4853:      <NA>         TRUE     <NA>               TRUE                 TRUE
#> 4854:      <NA>        FALSE     <NA>               TRUE                FALSE
#> 4855:      <NA>        FALSE     <NA>               TRUE                FALSE
#>       DateOriginal
#>    1:   2016-09-14
#>    2:   2016-09-14
#>    3:   2016-09-14
#>    4:   2016-09-14
#>    5:   2016-09-14
#>   ---             
#> 4851:   2020-06-04
#> 4852:   2020-06-04
#> 4853:   2020-06-04
#> 4854:   2020-06-04
#> 4855:   2020-06-04
TestData[IdTree == "101433", LifeStatus := c(FALSE, TRUE, TRUE, TRUE, TRUE)]
#>          Site Plot Subplot PlotLat PlotLon XPlotUTM YPlotUTM PlotArea
#>    1: Paracou    6    none      NA      NA       NA       NA     6.25
#>    2: Paracou    6    none      NA      NA       NA       NA     6.25
#>    3: Paracou    6    none      NA      NA       NA       NA     6.25
#>    4: Paracou    6    none      NA      NA       NA       NA     6.25
#>    5: Paracou    6    none      NA      NA       NA       NA     6.25
#>   ---                                                                
#> 4851: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4852: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4853: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4854: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4855: Paracou    6    none      NA      NA       NA       NA     6.25
#>       PlotElevation SubplotLat SubplotLon XSubplotUTM YSubplotUTM SubplotArea
#>    1:            NA         NA         NA          NA          NA          NA
#>    2:            NA         NA         NA          NA          NA          NA
#>    3:            NA         NA         NA          NA          NA          NA
#>    4:            NA         NA         NA          NA          NA          NA
#>    5:            NA         NA         NA          NA          NA          NA
#>   ---                                                                        
#> 4851:            NA         NA         NA          NA          NA          NA
#> 4852:            NA         NA         NA          NA          NA          NA
#> 4853:            NA         NA         NA          NA          NA          NA
#> 4854:            NA         NA         NA          NA          NA          NA
#> 4855:            NA         NA         NA          NA          NA          NA
#>       MinDBH IdCensus Year Month  Day       Date TreeFieldNum IdTree
#>    1:  10.03     2016 2016  <NA> <NA> 2016-09-14           45 100665
#>    2:  10.03     2016 2016  <NA> <NA> 2016-09-14          148 100765
#>    3:  10.03     2016 2016  <NA> <NA> 2016-09-14           46 100666
#>    4:  10.03     2016 2016  <NA> <NA> 2016-09-14           49 100669
#>    5:  10.03     2016 2016  <NA> <NA> 2016-09-14            1 100621
#>   ---                                                               
#> 4851:  10.03     2020 2020  <NA> <NA> 2020-06-04          884 101479
#> 4852:  10.03     2020 2020  <NA> <NA> 2020-06-04          887 101482
#> 4853:  10.03     2020 2020  <NA> <NA> 2020-06-04          896 101491
#> 4854:  10.03     2020 2020  <NA> <NA> 2020-06-04          905 101500
#> 4855:  10.03     2020   NA  <NA> <NA> 2020-06-04          764 101362
#>       StemFieldNum        IdStem LifeStatus Diameter BD  Circ BCirc HOM POM
#>    1:         <NA> 100665_1_auto       TRUE    53.16 NA 167.0    NA  NA   0
#>    2:         <NA> 100765_1_auto       TRUE    14.32 NA  45.0    NA  NA   0
#>    3:         <NA> 100666_1_auto       TRUE    19.89 NA  62.5    NA  NA   0
#>    4:         <NA> 100669_1_auto       TRUE    41.38 NA 130.0    NA  NA   0
#>    5:         <NA> 100621_1_auto       TRUE    13.37 NA  42.0    NA  NA   0
#>   ---                                                                      
#> 4851:         <NA> 101479_1_auto       TRUE    16.39 NA  51.5    NA  NA   0
#> 4852:         <NA> 101482_1_auto       TRUE    20.85 NA  65.5    NA  NA   0
#> 4853:         <NA> 101491_1_auto       TRUE    29.60 NA  93.0    NA  NA   0
#> 4854:         <NA> 101500_1_auto       TRUE    27.53 NA  86.5    NA  NA   0
#> 4855:         <NA> 101362_1_auto       TRUE    27.85 NA  87.5    NA  NA   0
#>       BHOM BPOM TreeHeight TreeLat TreeLon XTreeUTM YTreeUTM XTreePlot
#>    1:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    2:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    3:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    4:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    5:   NA <NA>         NA      NA      NA       NA       NA        NA
#>   ---                                                                 
#> 4851:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4852:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4853:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4854:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4855:   NA <NA>         NA      NA      NA       NA       NA        NA
#>       YTreePlot XTreeSubplot YTreeSubplot        ScientificName   VernName
#>    1:        NA           NA           NA        Eperua falcata       wapa
#>    2:        NA           NA           NA Conceveiba guianensis     kusisi
#>    3:        NA           NA           NA    Moronobea coccinea  Moronobea
#>    4:        NA           NA           NA   Pouteria melanopoda    bakuman
#>    5:        NA           NA           NA Sandwithia guyanensis  wata tiki
#>   ---                                                                     
#> 4851:        NA           NA           NA   Lecythis persistens maho rouge
#> 4852:        NA           NA           NA   Lecythis persistens maho rouge
#> 4853:        NA           NA           NA        Eperua falcata       wapa
#> 4854:        NA           NA           NA          Licania alba       koko
#> 4855:        NA           NA           NA     Licania canescens       koko
#>                 Family      Genus    Species Subspecies Variety Voucher IdLevel
#>    1:         Fabaceae     Eperua    falcata       <NA>    <NA>    <NA>    <NA>
#>    2:    Euphorbiaceae Conceveiba guianensis       <NA>    <NA>    <NA>    <NA>
#>    3:       Clusiaceae  Moronobea   coccinea       <NA>    <NA>    <NA>    <NA>
#>    4:       Sapotaceae   Pouteria melanopoda       <NA>    <NA>    <NA>    <NA>
#>    5:    Euphorbiaceae Sandwithia guyanensis       <NA>    <NA>    <NA>    <NA>
#>   ---                                                                          
#> 4851:    Lecythidaceae   Lecythis persistens       <NA>    <NA>    <NA>    <NA>
#> 4852:    Lecythidaceae   Lecythis persistens       <NA>    <NA>    <NA>    <NA>
#> 4853:         Fabaceae     Eperua    falcata       <NA>    <NA>    <NA>    <NA>
#> 4854: Chrysobalanaceae    Licania       alba       <NA>    <NA>    <NA>    <NA>
#> 4855: Chrysobalanaceae    Licania  canescens       <NA>    <NA>    <NA>    <NA>
#>       Authority CommercialSp LifeForm LifeStatusOriginal CommercialSpOriginal
#>    1:      <NA>         TRUE     <NA>               TRUE                 TRUE
#>    2:      <NA>        FALSE     <NA>               TRUE                FALSE
#>    3:      <NA>         TRUE     <NA>               TRUE                 TRUE
#>    4:      <NA>        FALSE     <NA>               TRUE                FALSE
#>    5:      <NA>        FALSE     <NA>               TRUE                FALSE
#>   ---                                                                        
#> 4851:      <NA>        FALSE     <NA>               TRUE                FALSE
#> 4852:      <NA>        FALSE     <NA>               TRUE                FALSE
#> 4853:      <NA>         TRUE     <NA>               TRUE                 TRUE
#> 4854:      <NA>        FALSE     <NA>               TRUE                FALSE
#> 4855:      <NA>        FALSE     <NA>               TRUE                FALSE
#>       DateOriginal
#>    1:   2016-09-14
#>    2:   2016-09-14
#>    3:   2016-09-14
#>    4:   2016-09-14
#>    5:   2016-09-14
#>   ---             
#> 4851:   2020-06-04
#> 4852:   2020-06-04
#> 4853:   2020-06-04
#> 4854:   2020-06-04
#> 4855:   2020-06-04
TestData[IdTree == "101435", LifeStatus := c(TRUE, TRUE, NA, FALSE, TRUE)]
#>          Site Plot Subplot PlotLat PlotLon XPlotUTM YPlotUTM PlotArea
#>    1: Paracou    6    none      NA      NA       NA       NA     6.25
#>    2: Paracou    6    none      NA      NA       NA       NA     6.25
#>    3: Paracou    6    none      NA      NA       NA       NA     6.25
#>    4: Paracou    6    none      NA      NA       NA       NA     6.25
#>    5: Paracou    6    none      NA      NA       NA       NA     6.25
#>   ---                                                                
#> 4851: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4852: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4853: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4854: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4855: Paracou    6    none      NA      NA       NA       NA     6.25
#>       PlotElevation SubplotLat SubplotLon XSubplotUTM YSubplotUTM SubplotArea
#>    1:            NA         NA         NA          NA          NA          NA
#>    2:            NA         NA         NA          NA          NA          NA
#>    3:            NA         NA         NA          NA          NA          NA
#>    4:            NA         NA         NA          NA          NA          NA
#>    5:            NA         NA         NA          NA          NA          NA
#>   ---                                                                        
#> 4851:            NA         NA         NA          NA          NA          NA
#> 4852:            NA         NA         NA          NA          NA          NA
#> 4853:            NA         NA         NA          NA          NA          NA
#> 4854:            NA         NA         NA          NA          NA          NA
#> 4855:            NA         NA         NA          NA          NA          NA
#>       MinDBH IdCensus Year Month  Day       Date TreeFieldNum IdTree
#>    1:  10.03     2016 2016  <NA> <NA> 2016-09-14           45 100665
#>    2:  10.03     2016 2016  <NA> <NA> 2016-09-14          148 100765
#>    3:  10.03     2016 2016  <NA> <NA> 2016-09-14           46 100666
#>    4:  10.03     2016 2016  <NA> <NA> 2016-09-14           49 100669
#>    5:  10.03     2016 2016  <NA> <NA> 2016-09-14            1 100621
#>   ---                                                               
#> 4851:  10.03     2020 2020  <NA> <NA> 2020-06-04          884 101479
#> 4852:  10.03     2020 2020  <NA> <NA> 2020-06-04          887 101482
#> 4853:  10.03     2020 2020  <NA> <NA> 2020-06-04          896 101491
#> 4854:  10.03     2020 2020  <NA> <NA> 2020-06-04          905 101500
#> 4855:  10.03     2020   NA  <NA> <NA> 2020-06-04          764 101362
#>       StemFieldNum        IdStem LifeStatus Diameter BD  Circ BCirc HOM POM
#>    1:         <NA> 100665_1_auto       TRUE    53.16 NA 167.0    NA  NA   0
#>    2:         <NA> 100765_1_auto       TRUE    14.32 NA  45.0    NA  NA   0
#>    3:         <NA> 100666_1_auto       TRUE    19.89 NA  62.5    NA  NA   0
#>    4:         <NA> 100669_1_auto       TRUE    41.38 NA 130.0    NA  NA   0
#>    5:         <NA> 100621_1_auto       TRUE    13.37 NA  42.0    NA  NA   0
#>   ---                                                                      
#> 4851:         <NA> 101479_1_auto       TRUE    16.39 NA  51.5    NA  NA   0
#> 4852:         <NA> 101482_1_auto       TRUE    20.85 NA  65.5    NA  NA   0
#> 4853:         <NA> 101491_1_auto       TRUE    29.60 NA  93.0    NA  NA   0
#> 4854:         <NA> 101500_1_auto       TRUE    27.53 NA  86.5    NA  NA   0
#> 4855:         <NA> 101362_1_auto       TRUE    27.85 NA  87.5    NA  NA   0
#>       BHOM BPOM TreeHeight TreeLat TreeLon XTreeUTM YTreeUTM XTreePlot
#>    1:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    2:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    3:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    4:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    5:   NA <NA>         NA      NA      NA       NA       NA        NA
#>   ---                                                                 
#> 4851:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4852:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4853:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4854:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4855:   NA <NA>         NA      NA      NA       NA       NA        NA
#>       YTreePlot XTreeSubplot YTreeSubplot        ScientificName   VernName
#>    1:        NA           NA           NA        Eperua falcata       wapa
#>    2:        NA           NA           NA Conceveiba guianensis     kusisi
#>    3:        NA           NA           NA    Moronobea coccinea  Moronobea
#>    4:        NA           NA           NA   Pouteria melanopoda    bakuman
#>    5:        NA           NA           NA Sandwithia guyanensis  wata tiki
#>   ---                                                                     
#> 4851:        NA           NA           NA   Lecythis persistens maho rouge
#> 4852:        NA           NA           NA   Lecythis persistens maho rouge
#> 4853:        NA           NA           NA        Eperua falcata       wapa
#> 4854:        NA           NA           NA          Licania alba       koko
#> 4855:        NA           NA           NA     Licania canescens       koko
#>                 Family      Genus    Species Subspecies Variety Voucher IdLevel
#>    1:         Fabaceae     Eperua    falcata       <NA>    <NA>    <NA>    <NA>
#>    2:    Euphorbiaceae Conceveiba guianensis       <NA>    <NA>    <NA>    <NA>
#>    3:       Clusiaceae  Moronobea   coccinea       <NA>    <NA>    <NA>    <NA>
#>    4:       Sapotaceae   Pouteria melanopoda       <NA>    <NA>    <NA>    <NA>
#>    5:    Euphorbiaceae Sandwithia guyanensis       <NA>    <NA>    <NA>    <NA>
#>   ---                                                                          
#> 4851:    Lecythidaceae   Lecythis persistens       <NA>    <NA>    <NA>    <NA>
#> 4852:    Lecythidaceae   Lecythis persistens       <NA>    <NA>    <NA>    <NA>
#> 4853:         Fabaceae     Eperua    falcata       <NA>    <NA>    <NA>    <NA>
#> 4854: Chrysobalanaceae    Licania       alba       <NA>    <NA>    <NA>    <NA>
#> 4855: Chrysobalanaceae    Licania  canescens       <NA>    <NA>    <NA>    <NA>
#>       Authority CommercialSp LifeForm LifeStatusOriginal CommercialSpOriginal
#>    1:      <NA>         TRUE     <NA>               TRUE                 TRUE
#>    2:      <NA>        FALSE     <NA>               TRUE                FALSE
#>    3:      <NA>         TRUE     <NA>               TRUE                 TRUE
#>    4:      <NA>        FALSE     <NA>               TRUE                FALSE
#>    5:      <NA>        FALSE     <NA>               TRUE                FALSE
#>   ---                                                                        
#> 4851:      <NA>        FALSE     <NA>               TRUE                FALSE
#> 4852:      <NA>        FALSE     <NA>               TRUE                FALSE
#> 4853:      <NA>         TRUE     <NA>               TRUE                 TRUE
#> 4854:      <NA>        FALSE     <NA>               TRUE                FALSE
#> 4855:      <NA>        FALSE     <NA>               TRUE                FALSE
#>       DateOriginal
#>    1:   2016-09-14
#>    2:   2016-09-14
#>    3:   2016-09-14
#>    4:   2016-09-14
#>    5:   2016-09-14
#>   ---             
#> 4851:   2020-06-04
#> 4852:   2020-06-04
#> 4853:   2020-06-04
#> 4854:   2020-06-04
#> 4855:   2020-06-04
TestData[IdTree == "101436", LifeStatus := c(TRUE, NA, NA, FALSE, NA)]
#>          Site Plot Subplot PlotLat PlotLon XPlotUTM YPlotUTM PlotArea
#>    1: Paracou    6    none      NA      NA       NA       NA     6.25
#>    2: Paracou    6    none      NA      NA       NA       NA     6.25
#>    3: Paracou    6    none      NA      NA       NA       NA     6.25
#>    4: Paracou    6    none      NA      NA       NA       NA     6.25
#>    5: Paracou    6    none      NA      NA       NA       NA     6.25
#>   ---                                                                
#> 4851: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4852: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4853: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4854: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4855: Paracou    6    none      NA      NA       NA       NA     6.25
#>       PlotElevation SubplotLat SubplotLon XSubplotUTM YSubplotUTM SubplotArea
#>    1:            NA         NA         NA          NA          NA          NA
#>    2:            NA         NA         NA          NA          NA          NA
#>    3:            NA         NA         NA          NA          NA          NA
#>    4:            NA         NA         NA          NA          NA          NA
#>    5:            NA         NA         NA          NA          NA          NA
#>   ---                                                                        
#> 4851:            NA         NA         NA          NA          NA          NA
#> 4852:            NA         NA         NA          NA          NA          NA
#> 4853:            NA         NA         NA          NA          NA          NA
#> 4854:            NA         NA         NA          NA          NA          NA
#> 4855:            NA         NA         NA          NA          NA          NA
#>       MinDBH IdCensus Year Month  Day       Date TreeFieldNum IdTree
#>    1:  10.03     2016 2016  <NA> <NA> 2016-09-14           45 100665
#>    2:  10.03     2016 2016  <NA> <NA> 2016-09-14          148 100765
#>    3:  10.03     2016 2016  <NA> <NA> 2016-09-14           46 100666
#>    4:  10.03     2016 2016  <NA> <NA> 2016-09-14           49 100669
#>    5:  10.03     2016 2016  <NA> <NA> 2016-09-14            1 100621
#>   ---                                                               
#> 4851:  10.03     2020 2020  <NA> <NA> 2020-06-04          884 101479
#> 4852:  10.03     2020 2020  <NA> <NA> 2020-06-04          887 101482
#> 4853:  10.03     2020 2020  <NA> <NA> 2020-06-04          896 101491
#> 4854:  10.03     2020 2020  <NA> <NA> 2020-06-04          905 101500
#> 4855:  10.03     2020   NA  <NA> <NA> 2020-06-04          764 101362
#>       StemFieldNum        IdStem LifeStatus Diameter BD  Circ BCirc HOM POM
#>    1:         <NA> 100665_1_auto       TRUE    53.16 NA 167.0    NA  NA   0
#>    2:         <NA> 100765_1_auto       TRUE    14.32 NA  45.0    NA  NA   0
#>    3:         <NA> 100666_1_auto       TRUE    19.89 NA  62.5    NA  NA   0
#>    4:         <NA> 100669_1_auto       TRUE    41.38 NA 130.0    NA  NA   0
#>    5:         <NA> 100621_1_auto       TRUE    13.37 NA  42.0    NA  NA   0
#>   ---                                                                      
#> 4851:         <NA> 101479_1_auto       TRUE    16.39 NA  51.5    NA  NA   0
#> 4852:         <NA> 101482_1_auto       TRUE    20.85 NA  65.5    NA  NA   0
#> 4853:         <NA> 101491_1_auto       TRUE    29.60 NA  93.0    NA  NA   0
#> 4854:         <NA> 101500_1_auto       TRUE    27.53 NA  86.5    NA  NA   0
#> 4855:         <NA> 101362_1_auto       TRUE    27.85 NA  87.5    NA  NA   0
#>       BHOM BPOM TreeHeight TreeLat TreeLon XTreeUTM YTreeUTM XTreePlot
#>    1:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    2:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    3:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    4:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    5:   NA <NA>         NA      NA      NA       NA       NA        NA
#>   ---                                                                 
#> 4851:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4852:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4853:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4854:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4855:   NA <NA>         NA      NA      NA       NA       NA        NA
#>       YTreePlot XTreeSubplot YTreeSubplot        ScientificName   VernName
#>    1:        NA           NA           NA        Eperua falcata       wapa
#>    2:        NA           NA           NA Conceveiba guianensis     kusisi
#>    3:        NA           NA           NA    Moronobea coccinea  Moronobea
#>    4:        NA           NA           NA   Pouteria melanopoda    bakuman
#>    5:        NA           NA           NA Sandwithia guyanensis  wata tiki
#>   ---                                                                     
#> 4851:        NA           NA           NA   Lecythis persistens maho rouge
#> 4852:        NA           NA           NA   Lecythis persistens maho rouge
#> 4853:        NA           NA           NA        Eperua falcata       wapa
#> 4854:        NA           NA           NA          Licania alba       koko
#> 4855:        NA           NA           NA     Licania canescens       koko
#>                 Family      Genus    Species Subspecies Variety Voucher IdLevel
#>    1:         Fabaceae     Eperua    falcata       <NA>    <NA>    <NA>    <NA>
#>    2:    Euphorbiaceae Conceveiba guianensis       <NA>    <NA>    <NA>    <NA>
#>    3:       Clusiaceae  Moronobea   coccinea       <NA>    <NA>    <NA>    <NA>
#>    4:       Sapotaceae   Pouteria melanopoda       <NA>    <NA>    <NA>    <NA>
#>    5:    Euphorbiaceae Sandwithia guyanensis       <NA>    <NA>    <NA>    <NA>
#>   ---                                                                          
#> 4851:    Lecythidaceae   Lecythis persistens       <NA>    <NA>    <NA>    <NA>
#> 4852:    Lecythidaceae   Lecythis persistens       <NA>    <NA>    <NA>    <NA>
#> 4853:         Fabaceae     Eperua    falcata       <NA>    <NA>    <NA>    <NA>
#> 4854: Chrysobalanaceae    Licania       alba       <NA>    <NA>    <NA>    <NA>
#> 4855: Chrysobalanaceae    Licania  canescens       <NA>    <NA>    <NA>    <NA>
#>       Authority CommercialSp LifeForm LifeStatusOriginal CommercialSpOriginal
#>    1:      <NA>         TRUE     <NA>               TRUE                 TRUE
#>    2:      <NA>        FALSE     <NA>               TRUE                FALSE
#>    3:      <NA>         TRUE     <NA>               TRUE                 TRUE
#>    4:      <NA>        FALSE     <NA>               TRUE                FALSE
#>    5:      <NA>        FALSE     <NA>               TRUE                FALSE
#>   ---                                                                        
#> 4851:      <NA>        FALSE     <NA>               TRUE                FALSE
#> 4852:      <NA>        FALSE     <NA>               TRUE                FALSE
#> 4853:      <NA>         TRUE     <NA>               TRUE                 TRUE
#> 4854:      <NA>        FALSE     <NA>               TRUE                FALSE
#> 4855:      <NA>        FALSE     <NA>               TRUE                FALSE
#>       DateOriginal
#>    1:   2016-09-14
#>    2:   2016-09-14
#>    3:   2016-09-14
#>    4:   2016-09-14
#>    5:   2016-09-14
#>   ---             
#> 4851:   2020-06-04
#> 4852:   2020-06-04
#> 4853:   2020-06-04
#> 4854:   2020-06-04
#> 4855:   2020-06-04

Rslt <- StatusCorrection(TestData[IdTree %in% selection],
                         InvariantColumns = c("Site",
                                              "Genus",
                                              "Species",
                                              "Family",
                                              "ScientificName"))

LifeStatusCorrectionPlot(Rslt)