StatusCorrectionByTree

Detect errors or detect errors and correct the tree life status evolution over the censuses. Inspired by the code of Nino Page package (ForestData::correct_alive() and .correct_alive_tree())

Usage

StatusCorrectionByTree(
  DataTree,
  PlotCensuses,
  InvariantColumns = c("Site", "GenusCor", "SpeciesCor", "FamilyCor",
    "ScientificNameCor"),
  DeathConfirmation = 2,
  UseSize = FALSE,
  AddRowsForForgottenCensuses = TRUE,
  DetectOnly = FALSE,
  RemoveRBeforeAlive = FALSE,
  RemoveRAfterDeath = FALSE
)

Arguments

DataTree

A dataset corresponding to a single tree's (1 IdTree/IdStem) measurements (data.frame or data.table) The LifeStatus column must be coded as: - TRUE = alive, - FALSE = dead, - NA = unseen

PlotCensuses

Census years for the plot in which the tree is (numeric or integer)

InvariantColumns

Vector with the names of the columns that are supposed to have always the same value for each measurement of the same tree (character). It is recommended to use the columns that have already been corrected, such as the columns containing the corrected botanical information.

DeathConfirmation

Number of times (censuses) needed for an unseen tree to be considered dead (numeric)

UseSize

Use the size presence as a witness of the living status of the tree (logical)

AddRowsForForgottenCensuses

TRUE: adds rows for forgotten censuses between 2 'Alive', FALSE: does not add any rows (logical)

DetectOnly

TRUE: Only detect errors, FALSE: detect and correct errors (logical)

RemoveRBeforeAlive

Do you want to delete the rows about the tree before it was seen alive for the 1st time? (logical)

RemoveRAfterDeath

After correction do you want to delete the rows concerning the tree after its death? (logical)

Value

Fill the Comment column with error type informations. If DetectOnly = FALSE, add a LifeStatusCor column with the corrected tree life status.

Details

if UseSize : if Diameter != NA -> Alive

Dead > Alive -> NA

add rows for the forgotten censuses between 2 'Alive' if chosen

Alive > Dead/NA > Alive -> Alive

Alive > NA > Dead -> NA

Alive > Dead > NA -> Dead

Alive > NA > NA: if DeathConfirmation > unseens -> NA if DeathConfirmation =< unseens -> Dead

Examples

library(data.table)
data(TestData)
Data <- TestData

# Wanted seq: Dead > Alive > NA > Dead > Alive > NA > NA > Dead > NA

DataTree <- Data[IdTree == "101623"]

AddR <- DataTree[2:5] # the rows to copy
AddR[, Year := c(2012:2015)] # the rows to add
#>    MinDBH IdCensus Year Month Day       Date    Site Cluster Plot PlotViewID
#> 1:  10.03     2020 2012     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 2:  10.03     2017 2013     7  19 2017-07-19 Paracou    <NA>    6       <NA>
#> 3:  10.03     2019 2014     6   7 2019-06-07 Paracou    <NA>    6       <NA>
#> 4:  10.03     2018 2015     7   8 2018-07-08 Paracou    <NA>    6       <NA>
#>    Subplot PlotArea PlotElevation SubplotArea PlotLat PlotLon XPlotUTM YPlotUTM
#> 1:       1     6.25            NA           0      NA      NA       NA       NA
#> 2:       1     6.25            NA           0      NA      NA       NA       NA
#> 3:       1     6.25            NA           0      NA      NA       NA       NA
#> 4:       1     6.25            NA           0      NA      NA       NA       NA
#>    SubplotLat SubplotLon XSubplotUTM YSubplotUTM         ScientificName
#> 1:         NA         NA          NA          NA Recordoxylon speciosum
#> 2:         NA         NA          NA          NA Recordoxylon speciosum
#> 3:         NA         NA          NA          NA Recordoxylon speciosum
#> 4:         NA         NA          NA          NA Recordoxylon speciosum
#>        VernName   Family        Genus   Species Subspecies Variety Voucher
#> 1: wakapu gitin Fabaceae Recordoxylon speciosum       <NA>    <NA>    <NA>
#> 2: wakapu gitin Fabaceae Recordoxylon speciosum       <NA>    <NA>    <NA>
#> 3: wakapu gitin Fabaceae Recordoxylon speciosum       <NA>    <NA>    <NA>
#> 4: wakapu gitin Fabaceae Recordoxylon speciosum       <NA>    <NA>    <NA>
#>    IdLevel Authority CommercialSp LifeForm TreeFieldNum IdTree StemFieldNum
#> 1:    <NA>      <NA>         TRUE     tree         2100 101623         <NA>
#> 2:    <NA>      <NA>         TRUE     tree         2100 101623         <NA>
#> 3:    <NA>      <NA>         TRUE     tree         2100 101623         <NA>
#> 4:    <NA>      <NA>         TRUE     tree         2100 101623         <NA>
#>           IdStem TreeLat TreeLon XTreeUTM YTreeUTM XTreePlot YTreePlot
#> 1: 101623_1_auto      NA      NA       NA       NA        NA        NA
#> 2: 101623_1_auto      NA      NA       NA       NA        NA        NA
#> 3: 101623_1_auto      NA      NA       NA       NA        NA        NA
#> 4: 101623_1_auto      NA      NA       NA       NA        NA        NA
#>    XTreeSubplot YTreeSubplot LifeStatus Diameter BD Circ BCirc HOM  POM BHOM
#> 1:           NA           NA       TRUE    21.49 NA 67.5    NA 1.3 <NA>   NA
#> 2:           NA           NA       TRUE    20.69 NA 65.0    NA 1.3 <NA>   NA
#> 3:           NA           NA       TRUE    21.33 NA 67.0    NA 1.3 <NA>   NA
#> 4:           NA           NA       TRUE    21.01 NA 66.0    NA 1.3 <NA>   NA
#>    BPOM TreeHeight TreeFieldNumOriginal IdTreeOriginal CommercialSpOriginal
#> 1: <NA>         NA                 2100         101623                 TRUE
#> 2: <NA>         NA                 2100         101623                 TRUE
#> 3: <NA>         NA                 2100         101623                 TRUE
#> 4: <NA>         NA                 2100         101623                 TRUE
#>    DateOriginal LifeStatusOriginal
#> 1:   2020-06-04               TRUE
#> 2:   2017-07-19               TRUE
#> 3:   2019-06-07               TRUE
#> 4:   2018-07-08               TRUE

DataTree <- rbindlist(list(DataTree, AddR)) # add rows

DataTree <- DataTree[order(Year)] # arrange years in ascending order

# Write the sequence
DataTree[, LifeStatus := c(FALSE, TRUE, NA, FALSE, TRUE, NA, NA, FALSE, NA)]
#>    MinDBH IdCensus Year Month Day       Date    Site Cluster Plot PlotViewID
#> 1:  10.03     2020 2012     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#> 2:  10.03     2017 2013     7  19 2017-07-19 Paracou    <NA>    6       <NA>
#> 3:  10.03     2019 2014     6   7 2019-06-07 Paracou    <NA>    6       <NA>
#> 4:  10.03     2018 2015     7   8 2018-07-08 Paracou    <NA>    6       <NA>
#> 5:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#> 6:  10.03     2017 2017     7  19 2017-07-19 Paracou    <NA>    6       <NA>
#> 7:  10.03     2018 2018     7   8 2018-07-08 Paracou    <NA>    6       <NA>
#> 8:  10.03     2019 2019     6   7 2019-06-07 Paracou    <NA>    6       <NA>
#> 9:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#>    Subplot PlotArea PlotElevation SubplotArea PlotLat PlotLon XPlotUTM YPlotUTM
#> 1:       1     6.25            NA           0      NA      NA       NA       NA
#> 2:       1     6.25            NA           0      NA      NA       NA       NA
#> 3:       1     6.25            NA           0      NA      NA       NA       NA
#> 4:       1     6.25            NA           0      NA      NA       NA       NA
#> 5:       1     6.25            NA           0      NA      NA       NA       NA
#> 6:       1     6.25            NA           0      NA      NA       NA       NA
#> 7:       1     6.25            NA           0      NA      NA       NA       NA
#> 8:       1     6.25            NA           0      NA      NA       NA       NA
#> 9:       1     6.25            NA           0      NA      NA       NA       NA
#>    SubplotLat SubplotLon XSubplotUTM YSubplotUTM         ScientificName
#> 1:         NA         NA          NA          NA Recordoxylon speciosum
#> 2:         NA         NA          NA          NA Recordoxylon speciosum
#> 3:         NA         NA          NA          NA Recordoxylon speciosum
#> 4:         NA         NA          NA          NA Recordoxylon speciosum
#> 5:         NA         NA          NA          NA Recordoxylon speciosum
#> 6:         NA         NA          NA          NA Recordoxylon speciosum
#> 7:         NA         NA          NA          NA Recordoxylon speciosum
#> 8:         NA         NA          NA          NA Recordoxylon speciosum
#> 9:         NA         NA          NA          NA Recordoxylon speciosum
#>        VernName   Family        Genus   Species Subspecies Variety Voucher
#> 1: wakapu gitin Fabaceae Recordoxylon speciosum       <NA>    <NA>    <NA>
#> 2: wakapu gitin Fabaceae Recordoxylon speciosum       <NA>    <NA>    <NA>
#> 3: wakapu gitin Fabaceae Recordoxylon speciosum       <NA>    <NA>    <NA>
#> 4: wakapu gitin Fabaceae Recordoxylon speciosum       <NA>    <NA>    <NA>
#> 5: wakapu gitin Fabaceae Recordoxylon speciosum       <NA>    <NA>    <NA>
#> 6: wakapu gitin Fabaceae Recordoxylon speciosum       <NA>    <NA>    <NA>
#> 7: wakapu gitin Fabaceae Recordoxylon speciosum       <NA>    <NA>    <NA>
#> 8: wakapu gitin Fabaceae Recordoxylon speciosum       <NA>    <NA>    <NA>
#> 9: wakapu gitin Fabaceae Recordoxylon speciosum       <NA>    <NA>    <NA>
#>    IdLevel Authority CommercialSp LifeForm TreeFieldNum IdTree StemFieldNum
#> 1:    <NA>      <NA>         TRUE     tree         2100 101623         <NA>
#> 2:    <NA>      <NA>         TRUE     tree         2100 101623         <NA>
#> 3:    <NA>      <NA>         TRUE     tree         2100 101623         <NA>
#> 4:    <NA>      <NA>         TRUE     tree         2100 101623         <NA>
#> 5:    <NA>      <NA>         TRUE     tree         2100 101623         <NA>
#> 6:    <NA>      <NA>         TRUE     tree         2100 101623         <NA>
#> 7:    <NA>      <NA>         TRUE     tree         2100 101623         <NA>
#> 8:    <NA>      <NA>         TRUE     tree         2100 101623         <NA>
#> 9:    <NA>      <NA>         TRUE     tree         2100 101623         <NA>
#>           IdStem TreeLat TreeLon XTreeUTM YTreeUTM XTreePlot YTreePlot
#> 1: 101623_1_auto      NA      NA       NA       NA        NA        NA
#> 2: 101623_1_auto      NA      NA       NA       NA        NA        NA
#> 3: 101623_1_auto      NA      NA       NA       NA        NA        NA
#> 4: 101623_1_auto      NA      NA       NA       NA        NA        NA
#> 5: 101623_1_auto      NA      NA       NA       NA        NA        NA
#> 6: 101623_1_auto      NA      NA       NA       NA        NA        NA
#> 7: 101623_1_auto      NA      NA       NA       NA        NA        NA
#> 8: 101623_1_auto      NA      NA       NA       NA        NA        NA
#> 9: 101623_1_auto      NA      NA       NA       NA        NA        NA
#>    XTreeSubplot YTreeSubplot LifeStatus Diameter BD Circ BCirc HOM  POM BHOM
#> 1:           NA           NA      FALSE    21.49 NA 67.5    NA 1.3 <NA>   NA
#> 2:           NA           NA       TRUE    20.69 NA 65.0    NA 1.3 <NA>   NA
#> 3:           NA           NA         NA    21.33 NA 67.0    NA 1.3 <NA>   NA
#> 4:           NA           NA      FALSE    21.01 NA 66.0    NA 1.3 <NA>   NA
#> 5:           NA           NA       TRUE    20.37 NA 64.0    NA 1.3 <NA>   NA
#> 6:           NA           NA         NA    20.69 NA 65.0    NA 1.3 <NA>   NA
#> 7:           NA           NA         NA    21.01 NA 66.0    NA 1.3 <NA>   NA
#> 8:           NA           NA      FALSE    21.33 NA 67.0    NA 1.3 <NA>   NA
#> 9:           NA           NA         NA    21.49 NA 67.5    NA 1.3 <NA>   NA
#>    BPOM TreeHeight TreeFieldNumOriginal IdTreeOriginal CommercialSpOriginal
#> 1: <NA>         NA                 2100         101623                 TRUE
#> 2: <NA>         NA                 2100         101623                 TRUE
#> 3: <NA>         NA                 2100         101623                 TRUE
#> 4: <NA>         NA                 2100         101623                 TRUE
#> 5: <NA>         NA                 2100         101623                 TRUE
#> 6: <NA>         NA                 2100         101623                 TRUE
#> 7: <NA>         NA                 2100         101623                 TRUE
#> 8: <NA>         NA                 2100         101623                 TRUE
#> 9: <NA>         NA                 2100         101623                 TRUE
#>    DateOriginal LifeStatusOriginal
#> 1:   2020-06-04               TRUE
#> 2:   2017-07-19               TRUE
#> 3:   2019-06-07               TRUE
#> 4:   2018-07-08               TRUE
#> 5:   2016-09-14               TRUE
#> 6:   2017-07-19               TRUE
#> 7:   2018-07-08               TRUE
#> 8:   2019-06-07               TRUE
#> 9:   2020-06-04               TRUE

Rslt <- StatusCorrectionByTree(DataTree, PlotCensuses = c(2011:2021),
                               InvariantColumns = c("Site",
                                                    "Genus",
                                                    "Species",
                                                    "Family",
                                                    "ScientificName"))
setnames(Rslt, "LifeStatusCor", "LifeStatus_TreeDataCor", skip_absent=TRUE)
LifeStatusCorrectionPlot(Rslt)