Correction of 2 diameter values

Correction of 2 diameter values

Usage

TwoValDiameterCor(
  DataTree,
  Data,
  DBHCor,
  Time,
  CorrectionType = NULL,
  PositiveGrowthThreshold = 5,
  NegativeGrowthThreshold = -2,
  DBHRange,
  MinIndividualNbr,
  OtherCrit,
  DetectOnly = FALSE
)

Arguments

DataTree

A dataset corresponding to a single tree/stem's (1 IdTree/IdStem) measurements (data.table) If 'phylogenetic hierarchical' correction is asked, the dataset must contain the columns:

IdStem or IdTree(character)

ScientificNameCor (character)

GenusCor (character)

FamilyCor (character)

Diameter (numeric)

Year (numeric)

Data

Complete dataset (data.table) The dataset must contain the columns:

IdStem or IdTree (character)

ScientificNameCor (character)

GenusCor (character)

FamilyCor (character)

Diameter (numeric)

Year (numeric)

DBHCor

Diameter vector in cm (numeric)

Time

Time vector in years (numeric)

CorrectionType

Possible values:

NULL: The 1st non-NA diameter value is assigned to the 2nd non-NA diameter value

"phylogenetic hierarchical": replace abnormal growth with the average growth of other trees in the dataset, at the specific, genus, family or stand level, within a DBH range of x cm (DBHRange argument). If the number of these trees < n (MinIndividualNbr argument) at the specific level, we switch to the genus level etc.

PositiveGrowthThreshold

in cm/year : a tree widening by more than x cm/year is considered abnormal (numeric, 1 value)

NegativeGrowthThreshold

in cm/census : The possible positive measurement error (+n) cannot be corrected until the growth appears abnormal, but a negative measurement error can be allowed until -n (a tree does not decrease). Thus the positive measurement error (+n) is "compensated". (numeric, 1 value)

DBHRange

DBH range in cm to take into account to select other trees in the dataset to apply "phylogenetic hierarchical" correction (Default: 10 cm) (numeric, 1 value)

MinIndividualNbr

Minimum number of individuals to take into account in "phylogenetic hierarchical" correction (Default: 5) (numeric, 1 value)

OtherCrit

Other criteria to select the individuals used for the calculation of the mean growth. Give the name of the column(s) for which the individuals must have the same value as the tree to correct (e.g. c("Plot", "Subplot")) (character)

DetectOnly

TRUE: Only detect errors, FALSE: detect and correct errors (Default: FALSE) (logical)

Value

List of 2 objects:

DBHCor (numeric vector): corrected by the method

DataTree (data.table): with the DiameterCorrectionMeth column filled with correction method.

Examples

library(data.table)
data(TestData)

DataTree <- TestData[IdTree %in% "100771"]
DataTree <- DataTree[order(Year)] # order de dt

DataTree$Diameter <- c(NA, 13, NA, 8, NA)

DBHCor <- DataTree$Diameter
Time <- DataTree$Year

TwoValDiameterCor(DataTree,
                  Data = TestData,
                  DBHCor, Time,
                  CorrectionType = "phylogenetic hierarchical",
                  DBHRange = 10,
                  MinIndividualNbr = 1, OtherCrit = NULL)
#> Warning: Invalid .internal.selfref detected and fixed by taking a (shallow) copy of the data.table so that := can add this new column by reference. At an earlier point, this data.table has been copied by R (or was created manually using structure() or similar). Avoid names<- and attr<- which in R currently (and oddly) may copy the whole data.table. Use set* syntax instead to avoid copying: ?set, ?setnames and ?setattr. If this message doesn't help, please report your use case to the data.table issue tracker so the root cause can be fixed or this message improved.
#> $DBHCor
#> [1]       NA 13.00000       NA 13.05938       NA
#> 
#> $DataTree
#>       Site Plot Subplot PlotLat PlotLon XPlotUTM YPlotUTM PlotArea
#> 1: Paracou    6    none      NA      NA       NA       NA     6.25
#> 2: Paracou    6    none      NA      NA       NA       NA     6.25
#> 3: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4: Paracou    6    none      NA      NA       NA       NA     6.25
#> 5: Paracou    6    none      NA      NA       NA       NA     6.25
#>    PlotElevation SubplotLat SubplotLon XSubplotUTM YSubplotUTM SubplotArea
#> 1:            NA         NA         NA          NA          NA          NA
#> 2:            NA         NA         NA          NA          NA          NA
#> 3:            NA         NA         NA          NA          NA          NA
#> 4:            NA         NA         NA          NA          NA          NA
#> 5:            NA         NA         NA          NA          NA          NA
#>    MinDBH IdCensus Year Month  Day       Date TreeFieldNum IdTree StemFieldNum
#> 1:  10.03     2016 2016  <NA> <NA> 2016-09-14          154 100771         <NA>
#> 2:  10.03     2017 2017  <NA> <NA> 2017-07-19          154 100771         <NA>
#> 3:  10.03     2018 2018  <NA> <NA> 2018-07-08          154 100771         <NA>
#> 4:  10.03     2019 2019  <NA> <NA> 2019-06-07          154 100771         <NA>
#> 5:  10.03     2020 2020  <NA> <NA> 2020-06-04          154 100771         <NA>
#>           IdStem LifeStatus Diameter BD Circ BCirc HOM POM BHOM BPOM TreeHeight
#> 1: 100771_1_auto       TRUE       NA NA 94.5    NA  NA   0   NA <NA>         NA
#> 2: 100771_1_auto       TRUE       13 NA 94.5    NA  NA   0   NA <NA>         NA
#> 3: 100771_1_auto       TRUE       NA NA 94.5    NA  NA   0   NA <NA>         NA
#> 4: 100771_1_auto       TRUE        8 NA 94.5    NA  NA   0   NA <NA>         NA
#> 5: 100771_1_auto       TRUE       NA NA 94.5    NA  NA   0   NA <NA>         NA
#>    TreeLat TreeLon XTreeUTM YTreeUTM XTreePlot YTreePlot XTreeSubplot
#> 1:      NA      NA       NA       NA        NA        NA           NA
#> 2:      NA      NA       NA       NA        NA        NA           NA
#> 3:      NA      NA       NA       NA        NA        NA           NA
#> 4:      NA      NA       NA       NA        NA        NA           NA
#> 5:      NA      NA       NA       NA        NA        NA           NA
#>    YTreeSubplot         ScientificName  VernName        Family       Genus
#> 1:           NA Eschweilera decolorans maho noir Lecythidaceae Eschweilera
#> 2:           NA Eschweilera decolorans maho noir Lecythidaceae Eschweilera
#> 3:           NA Eschweilera decolorans maho noir Lecythidaceae Eschweilera
#> 4:           NA Eschweilera decolorans maho noir Lecythidaceae Eschweilera
#> 5:           NA Eschweilera decolorans maho noir Lecythidaceae Eschweilera
#>       Species Subspecies Variety Voucher IdLevel Authority CommercialSp
#> 1: decolorans       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE
#> 2: decolorans       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE
#> 3: decolorans       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE
#> 4: decolorans       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE
#> 5: decolorans       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE
#>    LifeForm LifeStatusOriginal CommercialSpOriginal DateOriginal
#> 1:     <NA>               TRUE                FALSE   2016-09-14
#> 2:     <NA>               TRUE                FALSE   2017-07-19
#> 3:     <NA>               TRUE                FALSE   2018-07-08
#> 4:     <NA>               TRUE                FALSE   2019-06-07
#> 5:     <NA>               TRUE                FALSE   2020-06-04
#>                    Comment DiameterCorrectionMeth   DBHCor
#> 1:                                                      NA
#> 2:                                                13.00000
#> 3:                                                      NA
#> 4: Abnormal diameter value                species 13.05938
#> 5:                                                      NA
#>