UniqueMeasurement

Of the repeated measurements of an individual in the same census year, keep only the measurements (rows) taken at the highest POM (KeepMeas = "MaxHOM"), and/or the most recent measurement (same year but more recent date) (KeepMeas = "MaxDate")

Usage

UniqueMeasurement(Data, KeepMeas = c("MaxHOM", "MaxDate"), ID = "IdStem")

Arguments

Data

Dataset (data.table)

KeepMeas

In case of multiple measurements in the same census year: Possible values: "MaxHOM", "MaxDate" (character).

"MaxHOM": keep the measurement taken at the highest HOM/POM

"MaxDate": keep the most recent measurement (same year but more recent date)

ID

Column name indicating the identifier of the individual (character)

Value

Dataset (data.table) with 1 measurement (1 row) per Year.

Examples

library(data.table)

Data <- data.table( # 7 rows
IdStem = c(rep("a", 4), rep("b", 2), "c"),
Year = c(rep(2000, 3), 2001, rep(2000, 3)),
Date = as.Date(c("2000-01-10", "2000-01-20", "2000-01-30", "2001-01-10",
                 rep("2000-01-10", 3))),
HOM = c(rep(1, 3), 2, 1, 2, 1)
)
Data
#>    IdStem Year       Date HOM
#> 1:      a 2000 2000-01-10   1
#> 2:      a 2000 2000-01-20   1
#> 3:      a 2000 2000-01-30   1
#> 4:      a 2001 2001-01-10   2
#> 5:      b 2000 2000-01-10   1
#> 6:      b 2000 2000-01-10   2
#> 7:      c 2000 2000-01-10   1

UniqueMeasurement(TestData)
#>       MinDBH IdCensus Year Month Day       Date    Site Cluster Plot PlotViewID
#>    1:  10.03     2019 2019     6   7 2019-06-07 Paracou    <NA>    6       <NA>
#>    2:  10.03     2018 2018     7   8 2018-07-08 Paracou    <NA>    6       <NA>
#>    3:  10.03     2020 2020     6   4 2020-06-04 Paracou    <NA>    6       <NA>
#>    4:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>    5:  10.03     2019 2019     6   7 2019-06-07 Paracou    <NA>    6       <NA>
#>   ---                                                                          
#> 4850:  10.03     2019 2019     6   7 2019-06-07 Paracou    <NA>    6       <NA>
#> 4851:  10.03     2017 2017     7  19 2017-07-19 Paracou    <NA>    6       <NA>
#> 4852:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#> 4853:  10.03     2018 2018     7   8 2018-07-08 Paracou    <NA>    6       <NA>
#> 4854:  10.03     2016 2016     9  14 2016-09-14 Paracou    <NA>    6       <NA>
#>       Subplot PlotArea PlotElevation SubplotArea PlotLat PlotLon XPlotUTM
#>    1:       1     6.25            NA           0      NA      NA       NA
#>    2:       1     6.25            NA           0      NA      NA       NA
#>    3:       1     6.25            NA           0      NA      NA       NA
#>    4:       1     6.25            NA           0      NA      NA       NA
#>    5:       1     6.25            NA           0      NA      NA       NA
#>   ---                                                                    
#> 4850:       1     6.25            NA           0      NA      NA       NA
#> 4851:       1     6.25            NA           0      NA      NA       NA
#> 4852:       1     6.25            NA           0      NA      NA       NA
#> 4853:       1     6.25            NA           0      NA      NA       NA
#> 4854:       1     6.25            NA           0      NA      NA       NA
#>       YPlotUTM SubplotLat SubplotLon XSubplotUTM YSubplotUTM
#>    1:       NA         NA         NA          NA          NA
#>    2:       NA         NA         NA          NA          NA
#>    3:       NA         NA         NA          NA          NA
#>    4:       NA         NA         NA          NA          NA
#>    5:       NA         NA         NA          NA          NA
#>   ---                                                       
#> 4850:       NA         NA         NA          NA          NA
#> 4851:       NA         NA         NA          NA          NA
#> 4852:       NA         NA         NA          NA          NA
#> 4853:       NA         NA         NA          NA          NA
#> 4854:       NA         NA         NA          NA          NA
#>                 ScientificName            VernName           Family
#>    1:      Licania membranacea          santi koko Chrysobalanaceae
#>    2:    Sandwithia guyanensis           wata tiki    Euphorbiaceae
#>    3:     Iryanthera sagotiana tossopassa montagne    Myristicaceae
#>    4:           Eperua falcata                wapa         Fabaceae
#>    5: Indet.Sapindaceae Indet.            tatu udu      Sapindaceae
#>   ---                                                              
#> 4850:             Licania alba                koko Chrysobalanaceae
#> 4851:     Vouacapoua americana             wacapou         Fabaceae
#> 4852:     Pachira dolichocalyx          maho coton        Malvaceae
#> 4853:       Pouteria fimbriata            niamboka       Sapotaceae
#> 4854:      Gustavia hexapetala              kusisi    Lecythidaceae
#>                   Genus      Species Subspecies Variety Voucher IdLevel
#>    1:           Licania  membranacea       <NA>    <NA>    <NA>    <NA>
#>    2:        Sandwithia   guyanensis       <NA>    <NA>    <NA>    <NA>
#>    3:        Iryanthera    sagotiana       <NA>    <NA>    <NA>    <NA>
#>    4:            Eperua      falcata       <NA>    <NA>    <NA>    <NA>
#>    5: Indet.Sapindaceae       Indet.       <NA>    <NA>    <NA>    <NA>
#>   ---                                                                  
#> 4850:           Licania         alba       <NA>    <NA>    <NA>    <NA>
#> 4851:        Vouacapoua    americana       <NA>    <NA>    <NA>    <NA>
#> 4852:           Pachira dolichocalyx       <NA>    <NA>    <NA>    <NA>
#> 4853:          Pouteria    fimbriata       <NA>    <NA>    <NA>    <NA>
#> 4854:          Gustavia   hexapetala       <NA>    <NA>    <NA>    <NA>
#>       Authority CommercialSp LifeForm TreeFieldNum IdTree StemFieldNum
#>    1:      <NA>        FALSE     tree           38 100658         <NA>
#>    2:      <NA>        FALSE     tree           39 100659         <NA>
#>    3:      <NA>         TRUE     tree           44 100664         <NA>
#>    4:      <NA>         TRUE     tree           45 100665         <NA>
#>    5:      <NA>        FALSE     tree           42 100662         <NA>
#>   ---                                                                 
#> 4850:      <NA>        FALSE     tree          916 101511         <NA>
#> 4851:      <NA>         TRUE     tree          918 101513         <NA>
#> 4852:      <NA>         TRUE     tree          923 101518         <NA>
#> 4853:      <NA>        FALSE     tree          929 101524         <NA>
#> 4854:      <NA>        FALSE     tree          933 101528         <NA>
#>              IdStem TreeLat TreeLon XTreeUTM YTreeUTM XTreePlot YTreePlot
#>    1: 100658_1_auto      NA      NA       NA       NA        NA        NA
#>    2: 100659_1_auto      NA      NA       NA       NA        NA        NA
#>    3: 100664_1_auto      NA      NA       NA       NA        NA        NA
#>    4: 100665_1_auto      NA      NA       NA       NA        NA        NA
#>    5: 100662_1_auto      NA      NA       NA       NA        NA        NA
#>   ---                                                                    
#> 4850: 101511_1_auto      NA      NA       NA       NA        NA        NA
#> 4851: 101513_1_auto      NA      NA       NA       NA        NA        NA
#> 4852: 101518_1_auto      NA      NA       NA       NA        NA        NA
#> 4853: 101524_1_auto      NA      NA       NA       NA        NA        NA
#> 4854: 101528_1_auto      NA      NA       NA       NA        NA        NA
#>       XTreeSubplot YTreeSubplot LifeStatus Diameter BD  Circ BCirc HOM  POM
#>    1:           NA           NA       TRUE    17.83 NA  56.0    NA 1.3 <NA>
#>    2:           NA           NA       TRUE    12.25 NA  38.5    NA 1.3 <NA>
#>    3:           NA           NA       TRUE    15.92 NA  50.0    NA 1.3 <NA>
#>    4:           NA           NA       TRUE    53.16 NA 167.0    NA 1.3 <NA>
#>    5:           NA           NA       TRUE    19.10 NA  60.0    NA 1.3 <NA>
#>   ---                                                                      
#> 4850:           NA           NA       TRUE    22.12 NA  69.5    NA 1.3 <NA>
#> 4851:           NA           NA       TRUE    35.01 NA 110.0    NA 1.3 <NA>
#> 4852:           NA           NA       TRUE    16.87 NA  53.0    NA 1.3 <NA>
#> 4853:           NA           NA       TRUE    18.62 NA  58.5    NA 1.3 <NA>
#> 4854:           NA           NA       TRUE    14.80 NA  46.5    NA 1.3 <NA>
#>       BHOM BPOM TreeHeight TreeFieldNumOriginal IdTreeOriginal
#>    1:   NA <NA>         NA                   38         100658
#>    2:   NA <NA>         NA                   39         100659
#>    3:   NA <NA>         NA                   44         100664
#>    4:   NA <NA>         NA                   45         100665
#>    5:   NA <NA>         NA                   42         100662
#>   ---                                                         
#> 4850:   NA <NA>         NA                  916         101511
#> 4851:   NA <NA>         NA                  918         101513
#> 4852:   NA <NA>         NA                  923         101518
#> 4853:   NA <NA>         NA                  929         101524
#> 4854:   NA <NA>         NA                  933         101528
#>       CommercialSpOriginal DateOriginal LifeStatusOriginal
#>    1:                FALSE   2019-06-07               TRUE
#>    2:                FALSE   2018-07-08               TRUE
#>    3:                 TRUE   2020-06-04               TRUE
#>    4:                 TRUE   2016-09-14               TRUE
#>    5:                FALSE   2019-06-07               TRUE
#>   ---                                                     
#> 4850:                FALSE   2019-06-07               TRUE
#> 4851:                 TRUE   2017-07-19               TRUE
#> 4852:                 TRUE   2016-09-14               TRUE
#> 4853:                FALSE   2018-07-08               TRUE
#> 4854:                FALSE   2016-09-14               TRUE