UniqueMeasurement

Of the repeated measurements of an individual in the same census year, keep only the measurements (rows) taken at the highest POM (KeepMeas = "MaxHOM"), and/or the most recent measurement (same year but more recent date) (KeepMeas = "MaxDate")

Usage

UniqueMeasurement(Data, KeepMeas = c("MaxHOM", "MaxDate"), ID = "IdStem")

Arguments

Data

Dataset (data.table)

KeepMeas

In case of multiple measurements in the same census year: Possible values: "MaxHOM", "MaxDate" (character).

"MaxHOM": keep the measurement taken at the highest HOM/POM

"MaxDate": keep the most recent measurement (same year but more recent date)

ID

Column name indicating the identifier of the individual (character)

Value

Dataset (data.table) with 1 measurement (1 row) per Year.

Examples

library(data.table)

Data <- data.table( # 7 rows
IdStem = c(rep("a", 4), rep("b", 2), "c"),
Year = c(rep(2000, 3), 2001, rep(2000, 3)),
Date = as.Date(c("2000-01-10", "2000-01-20", "2000-01-30", "2001-01-10",
                 rep("2000-01-10", 3))),
HOM = c(rep(1, 3), 2, 1, 2, 1)
)
Data
#>    IdStem Year       Date HOM
#> 1:      a 2000 2000-01-10   1
#> 2:      a 2000 2000-01-20   1
#> 3:      a 2000 2000-01-30   1
#> 4:      a 2001 2001-01-10   2
#> 5:      b 2000 2000-01-10   1
#> 6:      b 2000 2000-01-10   2
#> 7:      c 2000 2000-01-10   1

UniqueMeasurement(TestData)
#>          Site Plot Subplot PlotLat PlotLon XPlotUTM YPlotUTM PlotArea
#>    1: Paracou    6    none      NA      NA       NA       NA     6.25
#>    2: Paracou    6    none      NA      NA       NA       NA     6.25
#>    3: Paracou    6    none      NA      NA       NA       NA     6.25
#>    4: Paracou    6    none      NA      NA       NA       NA     6.25
#>    5: Paracou    6    none      NA      NA       NA       NA     6.25
#>   ---                                                                
#> 4850: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4851: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4852: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4853: Paracou    6    none      NA      NA       NA       NA     6.25
#> 4854: Paracou    6    none      NA      NA       NA       NA     6.25
#>       PlotElevation SubplotLat SubplotLon XSubplotUTM YSubplotUTM SubplotArea
#>    1:            NA         NA         NA          NA          NA          NA
#>    2:            NA         NA         NA          NA          NA          NA
#>    3:            NA         NA         NA          NA          NA          NA
#>    4:            NA         NA         NA          NA          NA          NA
#>    5:            NA         NA         NA          NA          NA          NA
#>   ---                                                                        
#> 4850:            NA         NA         NA          NA          NA          NA
#> 4851:            NA         NA         NA          NA          NA          NA
#> 4852:            NA         NA         NA          NA          NA          NA
#> 4853:            NA         NA         NA          NA          NA          NA
#> 4854:            NA         NA         NA          NA          NA          NA
#>       MinDBH IdCensus Year Month  Day       Date TreeFieldNum IdTree
#>    1:  10.03     2019 2019  <NA> <NA> 2019-06-07           38 100658
#>    2:  10.03     2018 2018  <NA> <NA> 2018-07-08           39 100659
#>    3:  10.03     2020 2020  <NA> <NA> 2020-06-04           44 100664
#>    4:  10.03     2016 2016  <NA> <NA> 2016-09-14           45 100665
#>    5:  10.03     2019 2019  <NA> <NA> 2019-06-07           42 100662
#>   ---                                                               
#> 4850:  10.03     2019 2019  <NA> <NA> 2019-06-07          916 101511
#> 4851:  10.03     2017 2017  <NA> <NA> 2017-07-19          918 101513
#> 4852:  10.03     2016 2016  <NA> <NA> 2016-09-14          923 101518
#> 4853:  10.03     2018 2018  <NA> <NA> 2018-07-08          929 101524
#> 4854:  10.03     2016 2016  <NA> <NA> 2016-09-14          933 101528
#>       StemFieldNum        IdStem LifeStatus Diameter BD  Circ BCirc HOM POM
#>    1:         <NA> 100658_1_auto       TRUE    17.83 NA  56.0    NA  NA   0
#>    2:         <NA> 100659_1_auto       TRUE    12.25 NA  38.5    NA  NA   0
#>    3:         <NA> 100664_1_auto       TRUE    15.92 NA  50.0    NA  NA   0
#>    4:         <NA> 100665_1_auto       TRUE    53.16 NA 167.0    NA  NA   0
#>    5:         <NA> 100662_1_auto       TRUE    19.10 NA  60.0    NA  NA   0
#>   ---                                                                      
#> 4850:         <NA> 101511_1_auto       TRUE    22.12 NA  69.5    NA  NA   0
#> 4851:         <NA> 101513_1_auto       TRUE    35.01 NA 110.0    NA  NA   0
#> 4852:         <NA> 101518_1_auto       TRUE    16.87 NA  53.0    NA  NA   0
#> 4853:         <NA> 101524_1_auto       TRUE    18.62 NA  58.5    NA  NA   0
#> 4854:         <NA> 101528_1_auto       TRUE    14.80 NA  46.5    NA  NA   0
#>       BHOM BPOM TreeHeight TreeLat TreeLon XTreeUTM YTreeUTM XTreePlot
#>    1:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    2:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    3:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    4:   NA <NA>         NA      NA      NA       NA       NA        NA
#>    5:   NA <NA>         NA      NA      NA       NA       NA        NA
#>   ---                                                                 
#> 4850:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4851:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4852:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4853:   NA <NA>         NA      NA      NA       NA       NA        NA
#> 4854:   NA <NA>         NA      NA      NA       NA       NA        NA
#>       YTreePlot XTreeSubplot YTreeSubplot           ScientificName
#>    1:        NA           NA           NA      Licania membranacea
#>    2:        NA           NA           NA    Sandwithia guyanensis
#>    3:        NA           NA           NA     Iryanthera sagotiana
#>    4:        NA           NA           NA           Eperua falcata
#>    5:        NA           NA           NA Indet.Sapindaceae Indet.
#>   ---                                                             
#> 4850:        NA           NA           NA             Licania alba
#> 4851:        NA           NA           NA     Vouacapoua americana
#> 4852:        NA           NA           NA     Pachira dolichocalyx
#> 4853:        NA           NA           NA       Pouteria fimbriata
#> 4854:        NA           NA           NA      Gustavia hexapetala
#>                  VernName           Family             Genus      Species
#>    1:          santi koko Chrysobalanaceae           Licania  membranacea
#>    2:           wata tiki    Euphorbiaceae        Sandwithia   guyanensis
#>    3: tossopassa montagne    Myristicaceae        Iryanthera    sagotiana
#>    4:                wapa         Fabaceae            Eperua      falcata
#>    5:            tatu udu      Sapindaceae Indet.Sapindaceae       Indet.
#>   ---                                                                    
#> 4850:                koko Chrysobalanaceae           Licania         alba
#> 4851:             wacapou         Fabaceae        Vouacapoua    americana
#> 4852:          maho coton        Malvaceae           Pachira dolichocalyx
#> 4853:            niamboka       Sapotaceae          Pouteria    fimbriata
#> 4854:              kusisi    Lecythidaceae          Gustavia   hexapetala
#>       Subspecies Variety Voucher IdLevel Authority CommercialSp LifeForm
#>    1:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     <NA>
#>    2:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     <NA>
#>    3:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     <NA>
#>    4:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     <NA>
#>    5:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     <NA>
#>   ---                                                                   
#> 4850:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     <NA>
#> 4851:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     <NA>
#> 4852:       <NA>    <NA>    <NA>    <NA>      <NA>         TRUE     <NA>
#> 4853:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     <NA>
#> 4854:       <NA>    <NA>    <NA>    <NA>      <NA>        FALSE     <NA>
#>       LifeStatusOriginal CommercialSpOriginal DateOriginal
#>    1:               TRUE                FALSE   2019-06-07
#>    2:               TRUE                FALSE   2018-07-08
#>    3:               TRUE                 TRUE   2020-06-04
#>    4:               TRUE                 TRUE   2016-09-14
#>    5:               TRUE                FALSE   2019-06-07
#>   ---                                                     
#> 4850:               TRUE                FALSE   2019-06-07
#> 4851:               TRUE                 TRUE   2017-07-19
#> 4852:               TRUE                 TRUE   2016-09-14
#> 4853:               TRUE                FALSE   2018-07-08
#> 4854:               TRUE                FALSE   2016-09-14